+++ frontend/src/pages/data-product/components/data-product-tabs/team-tab/components/team-table/team-table.component.tsx
@@ -9,8 +9,6 @@ import { getDataProductUsersTableColumns } from '@/pages/data-product/components
 import { selectCurrentUser } from '@/store/features/auth/auth-slice.ts';
 import { useCheckAccessQuery } from '@/store/features/authorization/authorization-api-slice';
 import {
-    useDenyMembershipAccessMutation,
-    useGrantMembershipAccessMutation,
     useRemoveMembershipAccessMutation,
     useUpdateMembershipRoleMutation,
 } from '@/store/features/data-product-memberships/data-product-memberships-api-slice.ts';
@@ -20,6 +18,7 @@ import { AuthorizationAction } from '@/types/authorization/rbac-actions';
 import { DataProductMembershipRole, DataProductUserMembership } from '@/types/data-product-membership';
 import { UserContract } from '@/types/users';
 import { getDoesUserHaveAnyDataProductMembership } from '@/utils/data-product-user-role.helper.ts';
+import { usePendingActionHandlers } from '@/utils/pending-request.helper';
 
 import styles from './team-table.module.scss';
 
@@ -40,8 +39,7 @@ export function TeamTable({ isCurrentUserDataProductOwner, dataProductId, dataPr
     const [updateMembershipRole, { isLoading: isUpdatingMembershipRole }] = useUpdateMembershipRoleMutation();
     const [removeUserFromDataProduct, { isLoading: isRemovingUserFromDataProduct }] =
         useRemoveMembershipAccessMutation();
-    const [grantMembershipAccess] = useGrantMembershipAccessMutation();
-    const [denyMembershipAccess] = useDenyMembershipAccessMutation();
+    const { handleGrantAccessToDataProduct, handleDenyAccessToDataProduct } = usePendingActionHandlers();
 
     const { data: edit_access } = useCheckAccessQuery(
         {
@@ -93,30 +91,6 @@ export function TeamTable({ isCurrentUserDataProductOwner, dataProductId, dataPr
             }
         },
         [dataProduct, removeUserFromDataProduct, t],
-    );
-
-    const handleGrantAccessToDataProduct = useCallback(
-        async (membershipId: string) => {
-            try {
-                await grantMembershipAccess({ membershipId }).unwrap();
-                dispatchMessage({ content: t('User has been granted access to the data product'), type: 'success' });
-            } catch (_error) {
-                dispatchMessage({ content: t('Failed to grant user access to the data product'), type: 'error' });
-            }
-        },
-        [grantMembershipAccess, t],
-    );
-
-    const handleDenyAccessToDataProduct = useCallback(
-        async (membershipId: string) => {
-            try {
-                await denyMembershipAccess({ membershipId }).unwrap();
-                dispatchMessage({ content: t('User access to the data product has been denied'), type: 'success' });
-            } catch (_error) {
-                dispatchMessage({ content: t('Failed to deny user access to the data product'), type: 'error' });
-            }
-        },
-        [denyMembershipAccess, t],
     );
 
     const handleRoleChange = useCallback(
