+++ frontend/src/pages/dataset/components/dataset-tabs/data-output-tab/components/data-output-table/data-output-table.component.tsx
@@ -6,12 +6,10 @@ import { TABLE_SUBSECTION_PAGINATION } from '@/constants/table.constants.ts';
 import { useTablePagination } from '@/hooks/use-table-pagination.tsx';
 import { useCheckAccessQuery } from '@/store/features/authorization/authorization-api-slice.ts';
 import { useRemoveDataOutputDatasetLinkMutation } from '@/store/features/data-outputs-datasets/data-outputs-datasets-api-slice.ts';
-import { useApproveDataOutputLinkMutation } from '@/store/features/data-outputs-datasets/data-outputs-datasets-api-slice.ts';
-import { useRejectDataOutputLinkMutation } from '@/store/features/data-outputs-datasets/data-outputs-datasets-api-slice.ts';
 import { dispatchMessage } from '@/store/features/feedback/utils/dispatch-feedback.ts';
 import { AuthorizationAction } from '@/types/authorization/rbac-actions.ts';
-import type { DataOutputDatasetLinkRequest } from '@/types/data-output-dataset/data-output-dataset-link.contract.ts';
 import type { DataOutputLink } from '@/types/dataset';
+import { usePendingActionHandlers } from '@/utils/pending-request.helper.ts';
 
 import styles from './data-output-table.module.scss';
 import { getDatasetDataProductsColumns } from './data-output-table-columns.tsx';
@@ -26,8 +24,12 @@ type Props = {
 
 export function DataOutputTable({ isCurrentDatasetOwner, datasetId, dataOutputs, isLoading }: Props) {
     const { t } = useTranslation();
-    const [approveDataOutputLink, { isLoading: isApprovingLink }] = useApproveDataOutputLinkMutation();
-    const [rejectDataOutputLink, { isLoading: isRejectingLink }] = useRejectDataOutputLinkMutation();
+    const {
+        handleAcceptDataOutputDatasetLink,
+        handleRejectDataOutputDatasetLink,
+        isApprovingDataOutputLink,
+        isRejectingDataOutputLink,
+    } = usePendingActionHandlers();
     const [removeDatasetFromDataOutput, { isLoading: isRemovingDatasetFromDataProduct }] =
         useRemoveDataOutputDatasetLinkMutation();
 
@@ -78,42 +80,6 @@ export function DataOutputTable({ isCurrentDatasetOwner, datasetId, dataOutputs,
         [datasetId, removeDatasetFromDataOutput, t],
     );
 
-    const handleAcceptDataOutputDatasetLink = useCallback(
-        async (request: DataOutputDatasetLinkRequest) => {
-            try {
-                await approveDataOutputLink(request).unwrap();
-                dispatchMessage({
-                    content: t('Dataset request has been successfully approved'),
-                    type: 'success',
-                });
-            } catch (_error) {
-                dispatchMessage({
-                    content: t('Failed to approve data output dataset link'),
-                    type: 'error',
-                });
-            }
-        },
-        [approveDataOutputLink, t],
-    );
-
-    const handleRejectDataOutputDatasetLink = useCallback(
-        async (request: DataOutputDatasetLinkRequest) => {
-            try {
-                await rejectDataOutputLink(request).unwrap();
-                dispatchMessage({
-                    content: t('Dataset access request has been successfully rejected'),
-                    type: 'success',
-                });
-            } catch (_error) {
-                dispatchMessage({
-                    content: t('Failed to reject data output dataset link'),
-                    type: 'error',
-                });
-            }
-        },
-        [rejectDataOutputLink, t],
-    );
-
     const columns: TableColumnsType<DataOutputLink> = useMemo(() => {
         return getDatasetDataProductsColumns({
             onRemoveDataOutputDatasetLink: handleRemoveDatasetFromDataOutput,
@@ -121,7 +87,7 @@ export function DataOutputTable({ isCurrentDatasetOwner, datasetId, dataOutputs,
             isDisabled: !isCurrentDatasetOwner,
             canAcceptNew: canAcceptNew,
             canRevokeNew: canRevokeNew,
-            isLoading: isRemovingDatasetFromDataProduct || isRejectingLink || isApprovingLink,
+            isLoading: isRemovingDatasetFromDataProduct || isRejectingDataOutputLink || isApprovingDataOutputLink,
             isCurrentDatasetOwner,
             onRejectDataOutputDatasetLink: handleRejectDataOutputDatasetLink,
             onAcceptDataOutputDatasetLink: handleAcceptDataOutputDatasetLink,
@@ -131,8 +97,8 @@ export function DataOutputTable({ isCurrentDatasetOwner, datasetId, dataOutputs,
         t,
         isCurrentDatasetOwner,
         isRemovingDatasetFromDataProduct,
-        isRejectingLink,
-        isApprovingLink,
+        isRejectingDataOutputLink,
+        isApprovingDataOutputLink,
         handleRejectDataOutputDatasetLink,
         handleAcceptDataOutputDatasetLink,
         canAcceptNew,
